requirements.txt
textx
Django==1.11.2
coreapi==2.3.1
Markdown==2.6.8
django-filter==1.0.4
django-crispy-forms==1.6.1
django-guardian==1.4.8
djangorestframework==3.6.3
requests
